service/pyproject.toml
[project]
name = "Warp Zone Web Service"
version = ""
description = "web service component of Warp Zone"
authors = [
    {name = "Paul Bailey", email = "paul.bailey@cognitivespace.com"},
]
dependencies = [
    "django~=3.2",
    "whitenoise~=5.3",
    "Brotli~=1.0",
    "dj-database-url~=0.5",
    "psycopg2-binary~=2.9",
    "gunicorn~=20.1",
    "ipython~=7.25",
    "django-fernet-fields~=0.6",
    "kubernetes~=17.17",
    "PyYAML~=5.4",
    "loguru~=0.5",
    "django-json-widget~=1.1",
]
requires-python = ">=3.9"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/cognitive-space/warpzone"

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]

[tool.pdm.scripts]
_.env_file = ".env"
manage = "python manage.py"
req = "pdm export -o requirements.txt --without-hashes --prod"
